Running builtin tests from vscode (#395)
* adding tests, better assertions

* updated buffering options

* add another test

* added test ui

* don't use array if not needed

* add server/client

* server impl

* test node 20, 22

* updated promptfoo

* [genai] front matter update

* so

* retry after parsing

* better help to configure keys

* indent body request

* handle multiple scripts

* start viewer

* better handling of starting the prmpt

* fix cli

* updated commands

* more logging

* more docs

* more docs

* more docs updates

* updated docs, added javascript

* updated instructions

* updated landing page

+++ packages/vscode/src/dotenv.ts
@@ -1,7 +1,11 @@
 import * as vscode from "vscode"
 import { checkFileExists, readFileText, writeFile } from "./fs"
+import { parseModelIdentifier } from "genaiscript-core"
 
-export async function setupDotEnv(projectUri: vscode.Uri): Promise<string> {
+export async function setupDotEnv(
+    projectUri: vscode.Uri,
+    modelId?: string
+): Promise<string> {
     // update .gitignore file
     if (!(await checkFileExists(projectUri, ".gitignore")))
         await writeFile(projectUri, ".gitignore", ".env\n")
@@ -11,23 +15,23 @@ export async function setupDotEnv(projectUri: vscode.Uri): Promise<string> {
             await writeFile(projectUri, ".gitignore", content + "\n.env\n")
     }
 
+    const { provider } = parseModelIdentifier(modelId)
+    const key = `${provider.toUpperCase()}_API_KEY`
+    const keys = `
+# GenAIScript configuration (https://microsoft.github.io/genaiscript/reference/token/)
+${key}=<your token for ${provider}>
+#${provider.toUpperCase()}_API_BASE=${provider} api url
+`
+
     // update .env
     const uri = vscode.Uri.joinPath(projectUri, ".env")
-    if (!(await checkFileExists(uri)))
-        await writeFile(
-            projectUri,
-            ".env",
-            `# GenAIScript configuration (https://microsoft.github.io/genaiscript/reference/token/)
-OPENAI_API_KEY="<your token>"
-`
-        )
+    if (!(await checkFileExists(uri))) await writeFile(projectUri, ".env", keys)
 
     const doc = await vscode.workspace.openTextDocument(uri)
     await vscode.window.showTextDocument(doc)
     const text = doc.getText()
     let nextText = text
-    if (!/OPENAI_API_KEY/.test(text))
-        nextText += `\nOPENAI_API_KEY="<your token>"`
+    if (!text.includes(key)) nextText += `\n${keys}`
     if (nextText !== text) {
         const edit = new vscode.WorkspaceEdit()
         edit.replace(
